\name{hooah-package}
\alias{hooah-package}
\alias{hooah}
\docType{package}
\title{
R fragments.
}
\description{
My first R package.
}
\details{
\tabular{ll}{
Package: \tab hooah\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2006-12-21\cr
License: \tab GPL\cr
}
My first R package.
}
\author{
Zepu Zhang <zpzhang@stanfordalumnir.og>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
